refactor: 게임 초기 메시지 출력 및 참가자 카드 출력 메서드 분리
